separate RoomDetailsPage from switch logic
