+++ src/modules/page/room/RoomDetailsSwitch.page.vue
@@ -1,0 +1,43 @@
+<template>
+	<div v-if="isLoading" />
+	<template v-else>
+		<RoomDetailsPage v-if="isRoom" />
+		<CourseRoomDetailsPage v-else />
+	</template>
+</template>
+
+<script setup lang="ts">
+import CourseRoomDetailsPage from "@/pages/course-rooms/CourseRoomDetails.page.vue";
+import { RoomDetailsPage } from "@page-room";
+import { ENV_CONFIG_MODULE_KEY, injectStrict } from "@/utils/inject";
+import { RoomVariant, useRoomDetailsStore } from "@data-room";
+import { storeToRefs } from "pinia";
+import { computed, onUnmounted, watch } from "vue";
+import { useRoute } from "vue-router";
+
+const envConfigModule = injectStrict(ENV_CONFIG_MODULE_KEY);
+
+const route = useRoute();
+
+const roomDetailsStore = useRoomDetailsStore();
+const { isLoading, roomVariant } = storeToRefs(roomDetailsStore);
+const { deactivateRoom, fetchRoom, resetState } = roomDetailsStore;
+
+watch(
+	() => route.params.id,
+	async () => {
+		if (envConfigModule.getEnv["FEATURE_ROOMS_ENABLED"]) {
+			await fetchRoom(route.params.id as string);
+		} else {
+			deactivateRoom();
+		}
+	},
+	{ immediate: true }
+);
+
+const isRoom = computed(() => roomVariant.value === RoomVariant.ROOM);
+
+onUnmounted(() => {
+	resetState();
+});
+</script>
