+++ src/modules/page/room/index.ts
@@ -2,6 +2,7 @@ import RoomsPage from "./Rooms.page.vue";
 import RoomCreatePage from "./RoomCreate.page.vue";
 import RoomEditPage from "./RoomEdit.page.vue";
 import RoomDetailsPage from "./RoomDetails.page.vue";
+import RoomDetailsSwitchPage from "./RoomDetailsSwitch.page.vue";
 import RoomMembersPage from "./RoomMembers.page.vue";
 
 export {
@@ -9,5 +10,6 @@ export {
 	RoomCreatePage,
 	RoomEditPage,
 	RoomDetailsPage,
+	RoomDetailsSwitchPage,
 	RoomMembersPage,
 };
